working recursion with tests.
